mpeg4video: make i_modulo_time_base uint64_t
It's always increment with positive values, up to 64 bits.

(cherry picked from commit 8e6fc030bd650ce56c616e0cd220d1e777266bd0) (rebased)
rebased:
- the code around is slightly different

+++ modules/packetizer/mpeg4video.c
@@ -488,7 +488,7 @@ static int ParseVOP( decoder_t *p_dec, block_t *p_vop )
     decoder_sys_t *p_sys = p_dec->p_sys;
     int64_t i_time_ref;
     uint32_t i_time_increment;
-    int i_modulo_time_base = 0;
+    uint64_t i_modulo_time_base = 0;
     bs_t s;
 
     bs_init( &s, &p_vop->p_buffer[4], p_vop->i_buffer - 4 );
